+++ dart/lib/hct/hct.dart
@@ -15,7 +15,7 @@
 import 'package:material_color_utilities/utils/color_utils.dart';
 
 import 'cam16.dart';
-import 'cam_solver.dart';
+import 'hct_solver.dart';
 
 /// HCT, hue, chroma, and tone. A color system that provides a perceptually
 /// accurate color measurement system that can also accurately render what
@@ -32,7 +32,7 @@ class Hct {
   ///    given hue and tone.
   /// 0 <= [tone] <= 100; informally, lightness. Invalid values are corrected.
   static Hct from(double hue, double chroma, double tone) {
-    final argb = CamSolver.solveToInt(hue, chroma, tone);
+    final argb = HctSolver.solveToInt(hue, chroma, tone);
     return Hct._(argb);
   }
 
@@ -70,7 +70,7 @@ class Hct {
   /// representation. If the HCT color is outside of the sRGB gamut, chroma
   /// will decrease until it is inside the gamut.
   set hue(double newHue) {
-    _argb = CamSolver.solveToInt(newHue, chroma, tone);
+    _argb = HctSolver.solveToInt(newHue, chroma, tone);
     final cam16 = Cam16.fromInt(_argb);
     _hue = cam16.hue;
     _chroma = cam16.chroma;
@@ -87,7 +87,7 @@ class Hct {
   /// representation. If the HCT color is outside of the sRGB gamut, chroma
   /// will decrease until it is inside the gamut.
   set chroma(double newChroma) {
-    _argb = CamSolver.solveToInt(hue, newChroma, tone);
+    _argb = HctSolver.solveToInt(hue, newChroma, tone);
     final cam16 = Cam16.fromInt(_argb);
     _hue = cam16.hue;
     _chroma = cam16.chroma;
@@ -105,7 +105,7 @@ class Hct {
   /// representation. If the HCT color is outside of the sRGB gamut, chroma
   /// will decrease until it is inside the gamut.
   set tone(double newTone) {
-    _argb = CamSolver.solveToInt(hue, chroma, newTone);
+    _argb = HctSolver.solveToInt(hue, chroma, newTone);
     final cam16 = Cam16.fromInt(_argb);
     _hue = cam16.hue;
     _chroma = cam16.chroma;